requirements.txt
numpy>=1.26
sentence-transformers>=2.7
faiss-cpu>=1.8
rank-bm25>=0.2.2
pymupdf>=1.24
python-docx>=1.1
spacy>=3.7
langdetect>=1.0.9
elasticsearch>=8.13
transformers>=4.40
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
